chore: rename night mode to dark theme in app
