+++ docs/changelog.md
@@ -1413,7 +1413,7 @@ experience.
 ### Highlights for users (since 26.1.124 / 26.2.125)
 
 * Highlight colors for code blocks now match the webapp and
-  offer more contrast, especially in night mode.
+  offer more contrast, especially in dark theme.
 
 Plus, like every release, many other improvements for your Zulip
 experience.
