+++ src/common/__tests__/getStatusBarColor-test.js
@@ -1,17 +1,17 @@
 /* @flow strict-local */
 import { getStatusBarColor } from '../ZulipStatusBar';
 
-const themeNight = 'night';
+const themeDark = 'dark';
 const themeDefault = 'default';
 
 describe('getStatusBarColor', () => {
   test('returns specific color when given, regardless of theme', () => {
     expect(getStatusBarColor('#fff', themeDefault)).toEqual('#fff');
-    expect(getStatusBarColor('#fff', themeNight)).toEqual('#fff');
+    expect(getStatusBarColor('#fff', themeDark)).toEqual('#fff');
   });
 
   test('returns color according to theme for default case', () => {
     expect(getStatusBarColor(undefined, themeDefault)).toEqual('white');
-    expect(getStatusBarColor(undefined, themeNight)).toEqual('hsl(212, 28%, 18%)');
+    expect(getStatusBarColor(undefined, themeDark)).toEqual('hsl(212, 28%, 18%)');
   });
 });
