+++ src/settings/SettingsScreen.js
@@ -38,12 +38,12 @@ export default function SettingsScreen(props: Props): Node {
   const dispatch = useDispatch();
 
   const handleThemeChange = useCallback(() => {
-    dispatch(setGlobalSettings({ theme: theme === 'default' ? 'night' : 'default' }));
+    dispatch(setGlobalSettings({ theme: theme === 'default' ? 'dark' : 'default' }));
   }, [theme, dispatch]);
 
   return (
     <Screen title="Settings">
-      <SwitchRow label="Night mode" value={theme === 'night'} onValueChange={handleThemeChange} />
+      <SwitchRow label="Dark theme" value={theme === 'dark'} onValueChange={handleThemeChange} />
       <SwitchRow
         label="Open links with in-app browser"
         value={shouldUseInAppBrowser(browser)}
