+++ static/translations/messages_en.json
@@ -95,7 +95,7 @@
   "{realm} is running Zulip Server {serverVersion}, which is unsupported. Please contact your administrator about upgrading.": "{realm} is running Zulip Server {serverVersion}, which is unsupported. Please contact your administrator about upgrading.",
   "Learn more": "Learn more",
   "Settings": "Settings",
-  "Night mode": "Night mode",
+  "Dark theme": "Dark theme",
   "Open links with in-app browser": "Open links with in-app browser",
   "Language": "Language",
   "Arabic": "Arabic",
